[DI-1412] Address review feedback
