+++ package.json
@@ -24,7 +24,6 @@
     "dayjs": "^1.8.29",
     "https-proxy-agent": "^5.0.0",
     "jsonwebtoken": "^8.5.1",
-    "localtunnel": "2.0.2",
     "lodash": "^4.17.21",
     "q": "2.0.x",
     "qs": "^6.9.4",
@@ -48,19 +47,21 @@
     "nock": "^13.2.9",
     "node-mocks-http": "^1.8.1",
     "ts-jest": "^28.0.3",
-    "typescript": "^4.7.2"
+    "typescript": "^4.7.2",
+    "localtunnel": "2.0.2"
   },
   "scripts": {
-    "test": "npm run test:javascript",
+    "test": "npm run test:javascript -- spec/*",
     "test:javascript": "jest --coverage --detectOpenHandles",
     "test:typescript": "tsc -t es5 examples/typescript/example.ts --noEmit --strict",
     "jshint": "jshint lib/rest/** lib/base/** lib/http/**",
     "jscs": "eslint lib/base/**/**.js lib/http/**/**.js --fix",
     "check": "npm run jshint && npm run jscs",
-    "ci": "npm run test -- spec/* && npm run nsp",
+    "ci": "npm run test && npm run nsp",
     "jsdoc": "jsdoc -r lib -d docs",
     "nsp": "npm audit --production",
-    "cluster": "jest --testPathPattern='.*cluster_test.spec.ts$' --testNamePattern='^(?!\\bLocal Cluster Test\\b)' --coverage"
+    "cluster": "jest ./cluster_test.spec.ts --coverage",
+    "webhook-test": "jest ./webhook_cluster_test.spec.ts --coverage"
   },
   "files": [
     "lib",
